+++ ChangeLog.md
@@ -1,5 +1,10 @@
 # Revision history for reflex-fsnotify
 
+## 0.2.0.0
+
+* Deprecate `watchDirectory`
+* Add `watchDir` and `watchTree` corresponding to the functions in `System.FSNotify`
+
 ## 0.1.0.0
 
 * Initial release containing `watchDirectory`.